Challenge 1- I started of by creating an array of numbers(comprising of positive, negative and neutral integers) in conformity with the instruction above. Next, I declared my global variable and defined my function by calling the keyword(function), I assign a descriptive name to it and gave it a parameter- the array of numbers to sum.
In the first section of my function, I employed the use of the 'for' loop to iterate over the data in the array and only then did I use the length property to iterate through the array numbers before utilize that same property to sum up the elements.
In the final section of the function, I print this result in the console.

To conclude, I execute my function by calling its name with its appointed parameter.

Challenge 3- The task here was converting an array of temperatures in Celsius to Fahrenheit and returning a new array containing the converted temperatures.
The first step I took was to create an array of temperatures in Celsius.
The next thing I did was unpacking the values of this array individually, using the destructuring assignment syntax.
 Afterwards, I created a function with parameters, following the appropriate syntax.
Following that, I have my math operation in compliance with the conversion formula given by Anne Marie Helmenstine, Ph.D., from the link provided in the question.
Above, I individually applied the conversion formula to each of the elements in the array.
Also, I used the .toFixed method to round up the results to whole number units.
Then, I printed my result in the console alongside the unit for Fahrenheit and executed the function.
Finally, I created an array of temperatures in Fahrenheit containing the result of the function.

The result from both challenges were mathematically accurate and it was a welcomed test of my knowledge of javascript.